Исправление ошибок прошлого. Добавил файл в пакет ../homework. Переименовал класс с большой буквы. Теперь каждая переменная объявлена в отдельной строке. Вернул сбежавшие скобки. Сделал красивый вывод на русском.
